# Conflicts:
#	onnx/src/main/kotlin/org/jetbrains/kotlinx/dl/api/inference/onnx/ONNXModels.kt

+++ examples/src/main/java/LeNetClassic.java
@@ -52,9 +52,9 @@ public class LeNetClassic {
 
         try (Sequential lenet5Classic = Sequential.of(
                 new Input(new long[]{IMAGE_SIZE, IMAGE_SIZE, NUM_CHANNELS}, "x"),
-                new Conv2D(6, new long[]{5, 5}, new long[]{1, 1, 1, 1}, new long[]{1, 1, 1, 1}, Activations.Tanh, new GlorotNormal(SEED), new Zeros(), null, null, null, ConvPadding.SAME, true, "conv2d_1"),
+                new Conv2D(6, new int[]{5, 5}, new int[]{1, 1, 1, 1}, new int[]{1, 1, 1, 1}, Activations.Tanh, new GlorotNormal(SEED), new Zeros(), null, null, null, ConvPadding.SAME, true, "conv2d_1"),
                 new MaxPool2D(new int[]{1, 2, 2, 1}, new int[]{1, 2, 2, 1}, ConvPadding.VALID, "maxPool_1"),
-                new Conv2D(16, new long[]{5, 5}, new long[]{1, 1, 1, 1}, new long[]{1, 1, 1, 1}, Activations.Tanh, new GlorotNormal(SEED), new Zeros(), null, null, null, ConvPadding.SAME, true, "conv2d_2"),
+                new Conv2D(16, new int[]{5, 5}, new int[]{1, 1, 1, 1}, new int[]{1, 1, 1, 1}, Activations.Tanh, new GlorotNormal(SEED), new Zeros(), null, null, null, ConvPadding.SAME, true, "conv2d_2"),
                 new MaxPool2D(new int[]{1, 2, 2, 1}, new int[]{1, 2, 2, 1}, ConvPadding.VALID, "maxPool_2"),
                 new Flatten(), // 3136
                 new Dense(120, Activations.Tanh, new GlorotNormal(SEED), new Constant(0.1f), null, null, null, true, "dense_1"),
